+++ tests/asgi/test_query.py
@@ -2,7 +2,7 @@ from starlette.testclient import TestClient
 
 import strawberry
 from strawberry.asgi import GraphQL as BaseGraphQL
-from strawberry.types import ExecutionResult
+from strawberry.types import ExecutionResult, Info
 
 
 def test_simple_query(schema, test_client):
@@ -60,9 +60,48 @@ def test_root_value(schema, test_client):
     assert response.json() == {"data": {"rootName": "Query"}}
 
 
+def test_context_response():
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def something(self, info: Info) -> str:
+            r = info.context["response"]
+            r.raw_headers.append((b"x-bar", b"bar"))
+            return "foo"
+
+    schema = strawberry.Schema(query=Query)
+    app = BaseGraphQL(schema)
+
+    test_client = TestClient(app)
+    response = test_client.post("/", json={"query": "{ something }"})
+
+    assert response.status_code == 200
+    assert response.json() == {"data": {"something": "foo"}}
+    assert response.headers.get("x-bar") == "bar"
+
+
+def test_can_set_custom_status_code():
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def something(self, info: Info) -> str:
+            r = info.context["response"]
+            r.status_code = 418
+            return "foo"
+
+    schema = strawberry.Schema(query=Query)
+    app = BaseGraphQL(schema)
+
+    test_client = TestClient(app)
+    response = test_client.post("/", json={"query": "{ something }"})
+
+    assert response.status_code == 418
+    assert response.json() == {"data": {"something": "foo"}}
+
+
 def test_custom_context():
     class CustomGraphQL(BaseGraphQL):
-        async def get_context(self, request):
+        async def get_context(self, request, response):
             return {
                 "request": request,
                 "custom_context_value": "Hi!",
@@ -71,7 +110,7 @@ def test_custom_context():
     @strawberry.type
     class Query:
         @strawberry.field
-        def custom_context_value(self, info) -> str:
+        def custom_context_value(self, info: Info) -> str:
             return info.context["custom_context_value"]
 
     schema = strawberry.Schema(query=Query)
@@ -92,7 +131,7 @@ def test_custom_process_result():
     @strawberry.type
     class Query:
         @strawberry.field
-        def abc(self, info) -> str:
+        def abc(self) -> str:
             return "ABC"
 
     schema = strawberry.Schema(query=Query)