+++ tests/django/test_views.py
@@ -1,5 +1,5 @@
 import json
-from typing import Optional
+from typing import Any, Optional
 
 import pytest
 
@@ -10,7 +10,7 @@ from django.test.client import RequestFactory
 import strawberry
 from strawberry.django.views import GraphQLView as BaseGraphQLView
 from strawberry.permission import BasePermission
-from strawberry.types import ExecutionResult
+from strawberry.types import ExecutionResult, Info
 
 from .app.models import Example
 
@@ -18,7 +18,7 @@ from .app.models import Example
 class AlwaysFailPermission(BasePermission):
     message = "You are not authorized"
 
-    def has_permission(self, source, info):
+    def has_permission(self, source: Any, info: Info, **kwargs) -> bool:
         return False
 
 
@@ -27,12 +27,33 @@ class Query:
     hello: str = "strawberry"
 
     @strawberry.field(permission_classes=[AlwaysFailPermission])
-    def always_fail(self, info) -> Optional[str]:
+    def always_fail(self) -> Optional[str]:
         return "Hey"
 
     @strawberry.field
-    def example(self, info) -> str:
+    def example(self) -> str:
         return Example.objects.first().name
+
+
+@strawberry.type
+class GetRequestValueWithDotNotationQuery:
+    @strawberry.field
+    def get_request_value(self, info: Info) -> str:
+        return info.context.request
+
+
+@strawberry.type
+class GetRequestValueUsingGetQuery:
+    @strawberry.field
+    def get_request_value(self, info: Info) -> str:
+        return info.context.get("request")
+
+
+@strawberry.type
+class GetRequestValueQuery:
+    @strawberry.field
+    def get_request_value(self, info: Info) -> str:
+        return info.context["request"]
 
 
 schema = strawberry.Schema(query=Query)
@@ -98,6 +119,7 @@ def test_graphql_query():
     response = GraphQLView.as_view(schema=schema)(request)
     data = json.loads(response.content.decode())
 
+    assert response["content-type"] == "application/json"
     assert data["data"]["hello"] == "strawberry"
 
 
@@ -141,20 +163,41 @@ def test_returns_errors_and_data():
     assert data["errors"][0]["message"] == "You are not authorized"
 
 
+@pytest.mark.parametrize(
+    "query",
+    (
+        GetRequestValueWithDotNotationQuery,
+        GetRequestValueUsingGetQuery,
+        GetRequestValueQuery,
+    ),
+)
+def test_strawberry_django_context(query):
+    factory = RequestFactory()
+
+    schema = strawberry.Schema(query=query)
+
+    query = "{ getRequestValue }"
+    request = factory.post(
+        "/graphql/", {"query": query}, content_type="application/json"
+    )
+
+    response = GraphQLView.as_view(schema=schema)(request)
+    data = json.loads(response.content.decode())
+    assert response.status_code == 200
+    assert data["data"] == {"getRequestValue": "<WSGIRequest: POST '/graphql/'>"}
+
+
 def test_custom_context():
     class CustomGraphQLView(BaseGraphQLView):
-        def get_context(self, request):
-            return {
-                "request": request,
-                "custom_value": "Hi!",
-            }
+        def get_context(self, request, response):
+            return {"request": request, "custom_value": "Hi!"}
 
     factory = RequestFactory()
 
     @strawberry.type
     class Query:
         @strawberry.field
-        def custom_context_value(self, info) -> str:
+        def custom_context_value(self, info: Info) -> str:
             return info.context["custom_value"]
 
     schema = strawberry.Schema(query=Query)
@@ -181,7 +224,7 @@ def test_custom_process_result():
     @strawberry.type
     class Query:
         @strawberry.field
-        def abc(self, info) -> str:
+        def abc(self) -> str:
             return "ABC"
 
     schema = strawberry.Schema(query=Query)
@@ -196,3 +239,80 @@ def test_custom_process_result():
 
     assert response.status_code == 200
     assert data == {}
+
+
+def test_can_set_cookies():
+    factory = RequestFactory()
+
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def abc(self, info: Info) -> str:
+            info.context.response.set_cookie("fruit", "strawberry")
+
+            return "ABC"
+
+    schema = strawberry.Schema(query=Query)
+
+    query = "{ abc }"
+    request = factory.post(
+        "/graphql/", {"query": query}, content_type="application/json"
+    )
+
+    response = GraphQLView.as_view(schema=schema)(request)
+    data = json.loads(response.content.decode())
+
+    assert response.status_code == 200
+    assert response.cookies["fruit"].value == "strawberry"
+    assert data == {"data": {"abc": "ABC"}}
+
+
+def test_can_set_headers():
+    factory = RequestFactory()
+
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def abc(self, info: Info) -> str:
+            info.context.response["My-Header"] = "header value"
+
+            return "ABC"
+
+    schema = strawberry.Schema(query=Query)
+
+    query = "{ abc }"
+    request = factory.post(
+        "/graphql/", {"query": query}, content_type="application/json"
+    )
+
+    response = GraphQLView.as_view(schema=schema)(request)
+    data = json.loads(response.content.decode())
+
+    assert response.status_code == 200
+    assert response["my-header"] == "header value"
+    assert data == {"data": {"abc": "ABC"}}
+
+
+def test_can_change_status_code():
+    factory = RequestFactory()
+
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def abc(self, info: Info) -> str:
+            info.context.response.status_code = 418
+
+            return "ABC"
+
+    schema = strawberry.Schema(query=Query)
+
+    query = "{ abc }"
+    request = factory.post(
+        "/graphql/", {"query": query}, content_type="application/json"
+    )
+
+    response = GraphQLView.as_view(schema=schema)(request)
+    data = json.loads(response.content.decode())
+
+    assert response.status_code == 418
+    assert data == {"data": {"abc": "ABC"}}