+++ tests/schema/test_basic.py
@@ -4,7 +4,23 @@ from dataclasses import InitVar, dataclass
 from enum import Enum
 from typing import Optional
 
+import pytest
+
 import strawberry
+
+
+def test_raises_exception_with_unsupported_types():
+    class SomeType:
+        ...
+
+    @strawberry.type
+    class Query:
+        example: SomeType
+
+    with pytest.raises(
+        TypeError, match="Query fields cannot be resolved. Unexpected type '.*'"
+    ):
+        strawberry.Schema(query=Query)
 
 
 def test_basic_schema():
@@ -60,7 +76,7 @@ def test_does_camel_case_conversion():
     @strawberry.type
     class Query:
         @strawberry.field
-        def hello_world(self, info, query_param: str) -> str:
+        def hello_world(self, query_param: str) -> str:
             return query_param
 
     schema = strawberry.Schema(query=Query)
@@ -83,11 +99,11 @@ def test_can_rename_fields():
     @strawberry.type
     class Query:
         @strawberry.field
-        def hello(self, info) -> Hello:
+        def hello(self) -> Hello:
             return Hello("hi")
 
         @strawberry.field(name="example1")
-        def example(self, info, query_param: str) -> str:
+        def example(self, query_param: str) -> str:
             return query_param
 
     schema = strawberry.Schema(query=Query)
@@ -137,11 +153,11 @@ def test_field_description():
         a: str = strawberry.field(description="Example")
 
         @strawberry.field
-        def b(self, info, id: int) -> str:
+        def b(self, id: int) -> str:
             return "I'm a resolver"
 
         @strawberry.field(description="Example C")
-        def c(self, info, id: int) -> str:
+        def c(self, id: int) -> str:
             return "I'm a resolver"
 
     schema = strawberry.Schema(query=Query)
@@ -172,11 +188,11 @@ def test_field_deprecated_reason():
         a: str = strawberry.field(deprecation_reason="Deprecated A")
 
         @strawberry.field
-        def b(self, info, id: int) -> str:
+        def b(self, id: int) -> str:
             return "I'm a resolver"
 
         @strawberry.field(deprecation_reason="Deprecated B")
-        def c(self, info, id: int) -> str:
+        def c(self, id: int) -> str:
             return "I'm a resolver"
 
     schema = strawberry.Schema(query=Query)
@@ -281,7 +297,7 @@ def test_parent_class_fields_are_inherited():
         cheese: str = "swiss"
 
         @strawberry.field
-        def friend(self, info) -> str:
+        def friend(self) -> str:
             return "food"
 
     @strawberry.type
@@ -289,13 +305,13 @@ def test_parent_class_fields_are_inherited():
         cake: str = "made_in_switzerland"
 
         @strawberry.field
-        def hello_this_is(self, info) -> str:
+        def hello_this_is(self) -> str:
             return "patrick"
 
     @strawberry.type
     class Query:
         @strawberry.field
-        def example(self, info) -> Schema:
+        def example(self) -> Schema:
             return Schema()
 
     schema = strawberry.Schema(query=Query)
@@ -327,7 +343,7 @@ def test_can_return_compatible_type():
     @strawberry.type
     class Query:
         @strawberry.field
-        def assortment(self, info) -> Cheese:
+        def assortment(self) -> Cheese:
             return Example(name="Asiago")  # type: ignore
 
     schema = strawberry.Schema(query=Query)
@@ -353,7 +369,7 @@ def test_init_var():
     @strawberry.type
     class Query:
         @strawberry.field
-        def category(self, info) -> Category:
+        def category(self) -> Category:
             return Category(name="example", id="123")
 
     schema = strawberry.Schema(query=Query)
@@ -374,7 +390,7 @@ def test_nested_types():
     @strawberry.type
     class Query:
         @strawberry.field
-        def user(self, info) -> User:
+        def user(self) -> User:
             return User(name="Patrick")
 
     schema = strawberry.Schema(query=Query)