+++ tests/schema/types/test_datetime.py
@@ -23,7 +23,7 @@ def test_serialization(typing, instance, serialized):
     @strawberry.type
     class Query:
         @strawberry.field
-        def serialize(self, info) -> typing:
+        def serialize(self) -> typing:
             return instance
 
     schema = strawberry.Schema(Query)
@@ -59,7 +59,7 @@ def test_deserialization(typing, name, instance, serialized):
         deserialized = None
 
         @strawberry.field
-        def deserialize(self, info, arg: typing) -> bool:
+        def deserialize(self, arg: typing) -> bool:
             Query.deserialized = arg
             return True
 
@@ -92,7 +92,7 @@ def test_deserialization_with_parse_literal(typing, instance, serialized):
         deserialized = None
 
         @strawberry.field
-        def deserialize(self, info, arg: typing) -> bool:
+        def deserialize(self, arg: typing) -> bool:
             Query.deserialized = arg
             return True