+++ tests/types/test_permissions.py
@@ -1,12 +1,15 @@
+from typing import Any
+
 import strawberry
 from strawberry.permission import BasePermission
+from strawberry.types import Info
 
 
 def test_permission_classes_basic_fields():
     class IsAuthenticated(BasePermission):
         message = "User is not authenticated"
 
-        def has_permission(self, source, info):
+        def has_permission(self, source: Any, info: Info, **kwargs) -> bool:
             return False
 
     @strawberry.type
@@ -26,13 +29,13 @@ def test_permission_classes():
     class IsAuthenticated(BasePermission):
         message = "User is not authenticated"
 
-        def has_permission(self, source, info):
+        def has_permission(self, source: Any, info: Info, **kwargs) -> bool:
             return False
 
     @strawberry.type
     class Query:
         @strawberry.field(permission_classes=[IsAuthenticated])
-        def user(self, info) -> str:
+        def user(self) -> str:
             return "patrick"
 
     definition = Query._type_definition